Adapted the analyzer to only report unsafe LINQ operations for ConcurrentDictionary.
