+++ src/ParallelHelper.Plugin/ReleaseNotes.txt
@@ -1,6 +1,7 @@
 ﻿v2.1.0
 
 - New Analyzer: PH_P010 - Async Instead of Continuation
+- Adapted Analyzer: PH_B008 - Now only reports unsafe LINQ operations for ConcurrentDictionary
 
 
 ------------------
